// web/src/types/symbol.tsx
export interface Data {
  pair: string;
  price: number;
  price1dAgo: number;
  price1hAgo: number;
  price7dAgo: number;
  priceChange1d: number;
  priceChange1h: number;
  priceChange7d: number;
  priceBitstamp: number;
  priceKraken: number;
  quoteVolume: number;
  lowestPriceExchange: string;
  highestPriceExchange: string;
  lowestPrice: number;
  highestPrice: number;
  arbitrage: number;
}